Принимает цвет, теоретически должен и состояние. Необходимо доделать отсылку сообщений. Проблема - конвертация инта явы к байту си

+++ app/src/main/java/eugeny/n7/lampapp/DataReceivedListener.java
@@ -1,7 +1,8 @@
 package eugeny.n7.lampapp;
 
+import java.io.BufferedInputStream;
+
 public interface DataReceivedListener {
 
-    public void dataReceived(byte[] data);
-
+    public void dataReceived(int[] data);
 }
